requirements.txt
beautifulsoup4==4.9.3
binary==1.0.0
bs4==0.0.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.7
idna==2.10
numpy==1.21.0
opencv-contrib-python==4.5.2.54
peewee==3.14.4
Pillow==8.3.0
pony==0.7.14
psycopg2==2.9.1
pycparser==2.20
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.6
vk-api==11.9.4
